+++ Sources/SuperwallKit/Paywall/View Controller/PaywallViewController.swift
@@ -331,6 +331,23 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
     webView.scrollView.isScrollEnabled = paywall.isScrollEnabled
   }
 
+  func closeSafari(completion: (() -> Void)? = nil) {
+    guard
+      isSafariVCPresented,
+      let safariVC = presentedViewController as? SFSafariViewController
+    else {
+      completion?()
+      return
+    }
+    safariVC.dismiss(
+      animated: true,
+      completion: completion
+    )
+    // Must set this maually because programmatically dismissing the SafariVC doesn't call its
+    // delegate method where we set this.
+    isSafariVCPresented = false
+  }
+
   private func loadWebViewFromArchive(url: URL) {
     webView.loadFileURL(url, allowingReadAccessTo: url)
   }
@@ -691,7 +708,10 @@ extension PaywallViewController: PaywallMessageHandlerDelegate {
   }
 
   func presentSafariInApp(_ url: URL) {
-    guard UIApplication.shared.canOpenURL(url) else {
+    guard let sharedApplication = UIApplication.sharedApplication else {
+      return
+    }
+    guard sharedApplication.canOpenURL(url) else {
       Logger.debug(
         logLevel: .warn,
         scope: .paywallViewController,
@@ -708,7 +728,10 @@ extension PaywallViewController: PaywallMessageHandlerDelegate {
   }
 
   func presentSafariExternal(_ url: URL) {
-    UIApplication.shared.open(url)
+    guard let sharedApplication = UIApplication.sharedApplication else {
+      return
+    }
+    sharedApplication.open(url)
   }
 
   func openDeepLink(_ url: URL) {
@@ -717,7 +740,10 @@ extension PaywallViewController: PaywallMessageHandlerDelegate {
       closeReason: .systemLogic
     ) { [weak self] in
       self?.eventDidOccur(.openedDeepLink(url: url))
-      UIApplication.shared.open(url)
+      guard let sharedApplication = UIApplication.sharedApplication else {
+        return
+      }
+      sharedApplication.open(url)
     }
   }
 }