+++ Sources/SuperwallKit/StoreKit/Transactions/Purchasing/StoreKit 2/ProductPurchaserSK2.swift
@@ -96,7 +96,10 @@ final class ProductPurchaserSK2: Purchasing {
       let result: StoreKit.Product.PurchaseResult
 
       #if os(visionOS)
-      guard let scene = await UIApplication.shared.connectedScenes.first else {
+      guard let sharedApplication = UIApplication.sharedApplication else {
+        return .cancelled
+      }
+      guard let scene = await sharedApplication.connectedScenes.first else {
         return .cancelled
       }
       result = try await product.purchase(confirmIn: scene, options: options)